pyproject.toml
[project]
name = "lerobot-workspace"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "lerobot[aloha,dynamixel,feetech,intelrealsense,pusht]>=0.4.3",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[tool.uv]
dev-dependencies = []
